requirements.txt
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
PyMySQL==0.8.0
pytest==3.8.1
pytest-cov==2.5.1
retrying==1.3.3
six==1.11.0
pluggy==0.7.1
pyyaml==3.13
